Second Commit. Have libgen proxy up and design done
Libgen proxy has been made and now libgen.js isn't directly required.
The api is pretty basic and can consume search and new arrivals.

Search is done just by POSTing the object with right parameters to the
server in the way described by libgen.js to /explore.

Arrivals are done by performing a GET on /arrivals/count? .Essentially,
the count value is optional.

In addition, the basic design for the generic display page is done.
Search and sort has to be done, but otherwise things are looking good